Move failing tests in matrix test to individual file, update cypress config to enable video
Rename test files for order

Add more time to check for tweet

Try fixing issue with tests in different WP versions

Add timeout before click and additional step for lower version metabox

Reduce timeouts in tests

Temporarily remove trunk from matrix tests

Revert "Temporarily remove trunk from matrix tests"

This reverts commit 1a5576b32f5397b6d5133d5713d60080243db618.
