+++ tests/cypress/config.json
@@ -6,5 +6,5 @@
   "videosFolder": "tests/cypress/videos",
   "downloadsFolder": "tests/cypress/downloads",
   "supportFile": "tests/cypress/support/index.js",
-  "video": false
+  "video": true
 }
