+++ tests/cypress/integration/classic-editor-a.test.js
@@ -26,19 +26,3 @@ describe( 'Tests that new post is not tweeted when box is unchecked', () => {
 		cy.get( '#autoshare_for_twitter_metabox' ).contains( 'This post was not tweeted' );
 	} );
 } );
-
-describe( 'Tests that new post is tweeted when box is checked', () => {
-	it( 'Tests that new post is tweeted when box is checked', () => {
-		cy.visitAdminPage( 'post-new.php' );
-		let postTitle = getRandomText(8);
-		cy.get( '#title' ).type( 'Random Post Title' + postTitle );
-		cy.get( '#autoshare-for-twitter-enable' ).click();
-		cy.get( '#publish' ).click();
-
-		// Post-publish.
-		cy.get( '#autoshare_for_twitter_metabox', { timeout: 10000 } ).should( 'be.visible' );
-		cy.get( '#autoshare_for_twitter_metabox' ).contains( 'Tweeted on' );
-	} );
-} );
-
-
